Don't build the libminicontrol-inter.a library
Instead of building libminicontrol-inter.a library mark its sources as
sources for lib-minicontrol-*.so libraries.

Static libraries like the internal libminicontrol-inter.a should not be
used to build shared libraries like libminicontrol-monitor.so.0.0.1. To
build static libraries sources are compiled without -fPIC while for
shared libraries it is required to build with the -fPIC option.

Change-Id: I593152d3d0ad2cadc0d4279ffc8f8671783e74d9
Signed-off-by: Łukasz Stelmach <l.stelmach@samsung.com>